Changed prototype of ds1302_set_lines(). This way, no casts are necessary when calling it with boolean operations.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@22095 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/ide64.c
@@ -659,8 +659,8 @@ aborted_command:
             if ((kill_port & 2) == 0) {
                 break;
             }
-            ds1302_set_lines(ds1302_context, 1u, 0u, (BYTE)(value & 1u));
-            ds1302_set_lines(ds1302_context, 1u, 1u, (BYTE)(value & 1u));
+            ds1302_set_lines(ds1302_context, 1u, 0u, value & 1u);
+            ds1302_set_lines(ds1302_context, 1u, 1u, value & 1u);
             return;
         case 0xfb:
             if (((kill_port & 0x02) == 0) && (value & 0x02)) {
